Don't lock the screen when it's already locked.
The status of the screensaver wasn't taken into account when locking it.
Now we only lock it if it's not already locked. This way we don't get the
annoying error message stating that there was an error starting the
screensaver.

Part of the code is taken from the Chromium project. Adapted to fit our
purposes.