Fixed pausing idleness checks when fullscreen
For whatever reason, if user activity is simulated inside the slot that is connected to `KIdleTime::timeoutReached`, `KIdleTime` won't time out anymore. The patch fixes the issue by using a singleshot timer.

Closes https://github.com/lxqt/lxqt-powermanagement/issues/255

NOTE: Simulating user activity is needed in this case because a window may lose its fullscreen state without the user's interaction (as when a video ends).