Consider stable state of battery (#332)
`Solid::Battery::NoCharge` is interpreted as "Stable", not "Empty". "Empty" is used only when a completely empty battery is in the stable state (i.e., it isn't charging).

The "charge emblem" isn't shown in the stable state because there's no charging in this state. Moreover, the info we get is about the battery, not the adapter connection.

Closes https://github.com/lxqt/lxqt-powermanagement/issues/295